Basic running version with Outstanding_Amount column

+++ src/main/java/com/mfd/inventorytracking/orderbill/OrderBillingItem.java
@@ -1,14 +1,11 @@
 package com.mfd.inventorytracking.orderbill;
 
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleIntegerProperty;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
+import javafx.beans.property.*;
 
- class OrderBillingItem {
-	 StringProperty part=new SimpleStringProperty();
-	 IntegerProperty quantity=new SimpleIntegerProperty();
-	 IntegerProperty pricePerUnit=new SimpleIntegerProperty();
+class OrderBillingItem {
+	StringProperty  part         =new SimpleStringProperty();
+	IntegerProperty quantity     =new SimpleIntegerProperty();
+	FloatProperty   pricePerUnit =new SimpleFloatProperty();
 
 	  StringProperty partProperty() {
 		 return part;
@@ -18,7 +15,7 @@ import javafx.beans.property.StringProperty;
 		 return quantity;
 	 }
 
-	  IntegerProperty pricePerUnitProperty() {
+	  FloatProperty pricePerUnitProperty() {
 		 return pricePerUnit;
 	 }
